requirements.txt
openai>=1.12.0
jinja2>=3.1.3
weasyprint>=60.2
flask>=3.0.0
python-dotenv>=1.0.0
pillow>=10.0.0
requests>=2.31.0
python-telegram-bot>=20.7